refactor(runners): subagent.ts → subagent/runner.ts for symmetry
Matches the claude/ and pi/ subdirectory layout.
